Add things

+++ pipeworks/src/things/sprite.h
@@ -1,6 +1,7 @@
 #include <stdint.h>
 
 #include <pipeworks/things/sprite.h>
+
 #include "../thing.h" // Include private header; we're doing private things
 
 struct pw_sprite {
@@ -10,7 +11,7 @@ struct pw_sprite {
     float width;
     float height;
     uint32_t numFrames;
-    uint16_t frameWidth;
-    uint16_t frameHeight;
+    uint32_t frameWidth;
+    uint32_t frameHeight;
     uint32_t ***frames; // frames[id][x][y] (32-bit 0xAARRGGBB)
 };
